Factor out a Clock class
